requirements.txt
# Core Dependencies
# Telegram client
Telethon>=1.34.0

# Data handling
pandas>=2.0.0
python-dateutil>=2.8.2

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# GUI Framework
PySide6>=6.6.0

# Notifications
plyer>=2.1.0

# System tray and utilities
pywin32>=306; platform_system == "Windows"

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-qt>=4.2.0
black>=23.0.0
pylint>=2.17.0

# Packaging (for creating .exe)
pyinstaller>=6.0.0

# Optional: Icons
qtawesome>=1.2.3
